fix(core-components): lazy-load react-syntax-highlighter and dagre
The core-components barrel (export *) eagerly loads every component and
its transitive dependencies when any single export is imported. This
pulls in ~14.5 MB of unnecessary modules, with the two largest
offenders being react-syntax-highlighter (8.7 MB via CodeSnippet) and
@dagrejs/dagre + d3 (1.1 MB via DependencyGraph).

Split both components into a thin React.lazy wrapper and the actual
implementation. The wrapper is what the barrel exports — it defers
loading the heavy dependency until the component actually renders.
The public API is unchanged (verified via api-reports).
